Blob on the logged out home page is responsive
